% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BDS_Box_ObsChg.R
\name{BDS_Box_ObsChg}
\alias{BDS_Box_ObsChg}
\title{BDS_Box_ObsChg}
\usage{
BDS_Box_ObsChg(bdsdset, treatmentname, useshortnames = c(TRUE, FALSE),
  oldnames, newnames, usepopflag = c(TRUE, FALSE), popflag, testname,
  yaxislabel, selectedvisits, perpage, dignum, redoutliers = c(TRUE, FALSE),
  horizontallines = c(TRUE, FALSE), enterlimits = c(TRUE, FALSE), ANRLO,
  ANRHI, outputdirectory, filetype = c("PNG", "TIFF", "JPEG"), pixelwidth,
  pixelheight, outputfontsize, charttitle)
}
\arguments{
\item{bdsdset}{Dataset in the form of an R Data Frame}

\item{treatmentname}{Which treatment arm variable? e.g."TRTA" TRTA, TRTP, etc}

\item{useshortnames}{#Rename Treatment Arms? (if you wish to display shorter names).TRUE OR FALSE}

\item{oldnames}{Treatment Arms old names .e.g "Xanomeline Low Dose","Xanomeline High Dose"}

\item{newnames}{Treatment Arms new names e.g., "X-low", "X-High"}

\item{usepopflag}{subset on a population flag. TRUE OR FALSE}

\item{popflag}{value "SAFFL"}

\item{testname}{test or parameter to be analyzed e.g."DIABP"}

\item{yaxislabel}{labels for y axis}

\item{selectedvisits}{visit numbers to be analyzed e.g 0,2,4,6,8,12,16,20,24}

\item{perpage}{how many visits to display per page}

\item{dignum}{number of digits in table, standard deviation = dignum +1}

\item{redoutliers}{True or False.}

\item{horizontallines}{True or False.}

\item{enterlimits}{True or False.}

\item{ANRLO}{lower limit(s)}

\item{ANRHI}{upper limit(s)}

\item{outputdirectory}{set output file directory}

\item{filetype}{output file type - TIFF or JPEG or PNG}

\item{pixelwidth}{choose output file size: pixel width}

\item{pixelheight}{choose output file size: pixel height}

\item{outputfontsize}{choose output font size}

\item{charttitle}{Title for the chart}
}
\value{
PhUSE Figure 7.3 Box plot - Measurements and Change from Baseline by Analysis Timepoint, Visit and Treatment
}
\description{
Combine outputs from BDS_Box_Obs and BDS_Box_Chg into a single page
}

